revert stack size optimization

+++ src/movesort.cc
@@ -29,7 +29,7 @@ void History::clear()
 	std::memset(h, 0, sizeof(h));
 }
 
-int16_t History::get(const board::Position& B, move::move_t m) const
+int History::get(const board::Position& B, move::move_t m) const
 {
 	const int us = B.get_turn(), piece = B.get_piece_on(m.fsq()), tsq = m.tsq();
 	assert(!move::is_cop(B, m) && piece_ok(piece));
@@ -37,7 +37,7 @@ int16_t History::get(const board::Position& B, move::move_t m) const
 	return h[us][piece][tsq];
 }
 
-void History::add(const board::Position& B, move::move_t m, int16_t bonus)
+void History::add(const board::Position& B, move::move_t m, int bonus)
 {
 	const int us = B.get_turn(), piece = B.get_piece_on(m.fsq()), tsq = m.tsq();
 	assert(!move::is_cop(B, m) && piece_ok(piece));
